Add files via upload
This is a short JavaScript that accesses an online database and returns the data in it.  I use node.js with this script.  

This particular database is the City of Baltimore, MD crime statistics database.  Note that depending on how far into a year it is it may take some time to download all the data in this database.